@@ -1,80 +1,49 @@
 {
-  "name": "angular-io-example",
-  "version": "1.0.0",
-  "private": true,
-  "description": "Example project from an angular.io guide.",
+  "name": "angular-graph-ql",
+  "version": "0.0.0",
   "scripts": {
-    "lint": "tslint ./src/**/*.ts -t verbose",
-    "start": "ng serve --proxy-config proxy.conf.json",
-    "test": "ng test",
-    "build": "ng build --prod",
-    "e2e": "ng e2e",
-    "ng": "ng"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
+  "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.0",
-    "@angular/common": "^6.0.0",
-    "@angular/compiler": "^6.0.0",
-    "@angular/core": "^6.0.0",
-    "@angular/forms": "^6.0.0",
-    "@angular/http": "^6.0.0",
-    "@angular/platform-browser": "^6.0.0",
-    "@angular/platform-browser-dynamic": "^6.0.0",
-    "@angular/router": "^6.0.0",
-    "@angular/upgrade": "^6.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
-    "@types/jspdf": "^1.2.2",
+    "@angular/animations": "~12.2.0",
+    "@angular/common": "~12.2.0",
+    "@angular/compiler": "~12.2.0",
+    "@angular/core": "~12.2.0",
+    "@angular/forms": "~12.2.0",
+    "@angular/localize": "~12.2.0",
+    "@angular/platform-browser": "~12.2.0",
+    "@angular/platform-browser-dynamic": "~12.2.0",
+    "@angular/router": "~12.2.0",
+    "@apollo/client": "^3.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^10.0.0",
     "@types/underscore": "^1.8.9",
-    "angular-calendar": "^0.26.4",
-    "angular-in-memory-web-api": "^0.6.0",
-    "angularx-flatpickr": "^6.1.0",
-    "bootstrap": "^4.1.3",
-    "core-js": "^2.5.4",
-    "date-fns": "^1.29.0",
-    "flatpickr": "^4.5.2",
-    "graphql": "^15.6.0",
-    "html2canvas": "^1.0.0-alpha.12",
-    "less": "^3.9.0",
-    "material-icons": "^0.3.0",
-    "moment": "^2.22.2",
-    "moment-timezone": "^0.5.23",
-    "ng-multiselect-dropdown": "^0.3.4",
-    "ngx-color-picker": "^7.2.0",
-    "ngx-spinner": "^6.1.2",
-    "rxjs": "^6.0.0",
-    "underscore": "^1.9.1",
-    "zone.js": "^0.8.24"
+    "angular-calendar": "^0.28.28",
+    "apollo-angular": "^2.6.0",
+    "bootstrap": "^4.5.0",
+    "date-fns": "^2.16.1",
+    "graphql": "^15.0.0",
+    "moment": "^2.29.1",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.0",
-    "@angular/cli": "^6.0.0",
-    "@angular/compiler-cli": "^6.0.0",
-    "@angular/platform-server": "^6.0.0",
-    "@types/jasmine": "~2.8.0",
-    "@types/jasminewd2": "^2.0.4",
-    "@types/node": "^6.0.45",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-phantomjs-launcher": "^1.0.2",
-    "lodash": "^4.16.2",
-    "phantomjs-prebuilt": "^2.1.7",
-    "protractor": "~5.3.0",
-    "ts-node": "^5.0.1",
-    "tslint": "^5.9.1",
-    "typescript": "2.7.2"
-  },
-  "repository": {},
-  "-vs-binding": {
-    "AfterBuild": [
-      "build"
-    ]
+    "@angular-devkit/build-angular": "~12.2.7",
+    "@angular/cli": "~12.2.7",
+    "@angular/compiler-cli": "~12.2.0",
+    "@types/jasmine": "~3.8.0",
+    "@types/node": "^12.11.1",
+    "jasmine-core": "~3.8.0",
+    "karma": "~6.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "typescript": "~4.3.5"
   }
 }
